config.py
import os
from dotenv import load_dotenv
from datetime import datetime

from agent.utils import get_embed_model, get_llm_model
load_dotenv()
from llama_index.core import Settings

# - You must ALWAYS include citations and sources for all information provided
# CITATION REQUIREMENTS:
# - For web_search results: Include the NHS URL source and publication date which are returned by the tool
# - For rag results: Include the specific document title and author details
# - Format citations clearly at the end of each response or inline where relevant
# - Never provide information without proper attribution

# System prompts for each assistant
SYSTEM_PROMPTS = {
    "uk": """You are a supportive assistant for UK parents with autistic children. Respond in a compassionate, informative, and locally relevant manner.

IMPORTANT TOOL USAGE:
- If the 'web_search' tool is available, you MUST use it to get the data about the question and then provide a response based on the search results
- If the 'rag' tool is available, you MUST use it to access data from certified research papers and publications by passing queries to the tool and then providing a response based on the results
- If both tools are available, then provide a response based on results from both tools
- When using web_search results, include any image URLs found in the results to provide visual context and render the image url as markdown

Provide practical, evidence-based advice while maintaining empathy and understanding of the challenges UK parents face.

Today's date: {date}""".format(date=datetime.now().strftime("%Y-%m-%d %A")),

    "india": """You are a helpful assistant for Indian parents with autistic children. Offer advice and support tailored to Indian context and resources.

IMPORTANT TOOL USAGE:
- If the 'web_search' tool is available, you MUST use it to get the data about the question and then provide a response based on the search results
- If the 'rag' tool is available, you MUST use it to access data from certified research papers and publications by passing queries to the tool and then providing a response based on the results
- If both tools are available, then provide a response based on results from both tools
- When using web_search results, include any image URLs found in the results to provide visual context and render the image url as markdown

Provide culturally sensitive, practical advice while considering Indian healthcare systems, educational resources, and social contexts.

Today's date: {date}""".format(date=datetime.now().strftime("%Y-%m-%d %A"))
}

# API key envs
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
TAVILY_API_KEY = os.getenv("TAVILY_API_KEY")

# Directory for docs
DOCS_DIRECTORY = "./data/docs"
VECTOR_INDEX_DIR = "./data/vector_index"

CHUNK_SIZE = 2048
CHUNK_OVERLAP = 50

Settings.llm = get_llm_model(GOOGLE_API_KEY)
Settings.embed_model = get_embed_model(GOOGLE_API_KEY)
Settings.chunk_size = CHUNK_SIZE
Settings.chunk_overlap = CHUNK_OVERLAP
