requirements.txt
streamlit
llama-index
llama-index-llms-google-genai
llama-index-embeddings-google-genai
tavily-python
llama-index-readers-docling